feat: week 2 day 3 task
